refactor : baseline 엔티티 averageEventsPerHour 데이터 long형으로 변환

+++ backend/src/main/java/com/dajava/backend/domain/event/es/entity/AbusingBaseLine.java
@@ -33,7 +33,7 @@ public class AbusingBaseLine extends BaseTimeEntity {
 	private String pageUrl;
 
 	@Column(nullable = false)
-	private int averageEventsPerHour;
+	private long averageEventsPerHour;
 
 	@Column(nullable = false)
 	private double standardDeviation;
